# src/main/java/br/com/extratora/twelvekingdoms/dto/request/UpdateNotesRequest.java
package br.com.extratora.twelvekingdoms.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.NotNull;
import javax.validation.constraints.Size;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UpdateNotesRequest {
    @Size(max = 1000)
    @NotNull
    private String notes;
}