# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: 12 Kingdoms Backend
  version: V2
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /sheets:
    get:
      tags:
        - sheet-controller-impl
      summary: Retrieve sheet details
      operationId: sheetDetails
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SheetModel'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - sheet-controller-impl
      summary: Create a sheet
      operationId: createSheet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSheetRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /sheets/{sheetId}/updateNotes:
    patch:
      tags:
        - sheet-controller-impl
      summary: Update notes of sheet
      operationId: updateNotes
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotesRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /sheets/{sheetId}/updateDeathRolls:
    patch:
      tags:
        - sheet-controller-impl
      summary: Update death rolls of sheet
      operationId: updateDeathRolls
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeathRollsRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /sheets/{sheetId}/removeWound:
    patch:
      tags:
        - sheet-controller-impl
      summary: Remove wound from existing sheet
      operationId: removeWound
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /sheets/{sheetId}/addWound:
    patch:
      tags:
        - sheet-controller-impl
      summary: Add wound to existing sheet
      operationId: addWound
      parameters:
        - name: woundId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sheetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /sheets/{id}/levelUp:
    patch:
      tags:
        - sheet-controller-impl
      summary: Level up a sheet
      operationId: levelUp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
  /sheets/{id}/currentPoints:
    patch:
      tags:
        - sheet-controller-impl
      summary: "Update current physical, mental and heroism points of a sheet"
      operationId: updateCurrentPoints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSheetCurrentPointsRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /sheets/list:
    get:
      tags:
        - sheet-controller-impl
      summary: List sheets
      operationId: sheetList
      parameters:
        - name: currentPage
          in: query
          required: false
          schema:
            minimum: 0
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 5
        - name: sortDirection
          in: query
          required: false
          schema:
            type: string
        - name: sortField
          in: query
          required: false
          schema:
            type: string
        - name: usePlayerProfile
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: nameFilter
          in: query
          required: false
          schema:
            type: string
            default: ""
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SheetListResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
  /sheets/{id}:
    delete:
      tags:
        - sheet-controller-impl
      summary: Delete a sheet
      operationId: deleteSheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "403":
          description: Forbidden
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
    CreateSheetRequest:
      required:
        - backgroundId
        - celerity
        - cunning
        - intelligence
        - jobId
        - lineageId
        - name
        - tenacity
      type: object
      properties:
        name:
          maxLength: 20
          minLength: 5
          type: string
        lineageId:
          type: string
          format: uuid
        backgroundId:
          type: string
          format: uuid
        jobId:
          type: string
          format: uuid
        aptitudeList:
          type: array
          items:
            type: string
            format: uuid
        intelligence:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        cunning:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        tenacity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        celerity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        bond:
          maxLength: 200
          minLength: 0
          type: string
        motivation:
          maxLength: 200
          minLength: 0
          type: string
    IdResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      properties:
        errorList:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDto'
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    UpdateNotesRequest:
      required:
        - notes
      type: object
      properties:
        notes:
          maxLength: 1000
          minLength: 0
          type: string
    UpdateDeathRollsRequest:
      required:
        - deathRollBody
        - deathRollMind
        - deathRollSpirit
      type: object
      properties:
        deathRollBody:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollMind:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollSpirit:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
    UpdateSheetCurrentPointsRequest:
      type: object
      properties:
        mentalCurrent:
          type: integer
          format: int32
        physicalCurrent:
          type: integer
          format: int32
        heroismCurrent:
          type: integer
          format: int32
...
    WoundsModel:
      required:
        - description
        - diceRange
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        diceRange:
          type: string
        description:
          type: string
    AdvantageModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    AptitudeModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    BackgroundModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        physicalPoints:
          type: integer
          format: int32
        mentalPoints:
          type: integer
          format: int32
        advantage:
          $ref: '#/components/schemas/AdvantageModel'
    JobModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        mainAttribute:
          type: string
          enum:
            - INTELLIGENCE
            - CUNNING
            - CELERITY
            - TENACITY
        physicalPoints:
          type: integer
          format: int32
        mentalPoints:
          type: integer
          format: int32
        physicalPerLevel:
          type: integer
          format: int32
        mentalPerLevel:
          type: integer
          format: int32
        skills:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/SkillModel'
        aptitudes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AptitudeModel'
    LineageModel:
      required:
        - description
        - language
        - maturity
        - negativeTraitDescription
        - negativeTraitName
        - positiveTraitDescription
        - positiveTraitName
        - size
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        size:
          type: string
        language:
          type: string
        maturity:
          type: string
        positiveTraitName:
          type: string
        positiveTraitDescription:
          type: string
        negativeTraitName:
          type: string
        negativeTraitDescription:
          type: string
    SheetModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        intelligence:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        cunning:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        tenacity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        celerity:
          type: string
          enum:
            - D4
            - D6
            - D8
            - D12
            - D20
        name:
          type: string
        level:
          type: integer
          format: int32
        mentalCurrent:
          type: integer
          format: int32
        mentalTotal:
          type: integer
          format: int32
        physicalCurrent:
          type: integer
          format: int32
        physicalTotal:
          type: integer
          format: int32
        heroismCurrent:
          type: integer
          format: int32
        heroismTotal:
          type: integer
          format: int32
        bond:
          type: string
        motivation:
          type: string
        notes:
          type: string
        deathRollBody:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollMind:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        deathRollSpirit:
          type: string
          enum:
            - UNCHECKED
            - SUCCESS
            - FAILURE
        lineage:
          $ref: '#/components/schemas/LineageModel'
        background:
          $ref: '#/components/schemas/BackgroundModel'
        job:
          $ref: '#/components/schemas/JobModel'
        wound:
          $ref: '#/components/schemas/WoundsModel'
        aptitudes:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AptitudeModel'
        active:
          type: boolean
    SkillModel:
      required:
        - description
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        cost:
          type: integer
          format: int32
        energyType:
          type: string
          enum:
            - PHYSICAL
            - MENTAL
        skillType:
          type: string
          enum:
            - INTERPRETATION
            - COMBAT
            - BASIC
        range:
          type: integer
          format: int32
        skillLevel:
          type: integer
          format: int32
    BasicSheetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        level:
          type: integer
          format: int32
        lineage:
          type: string
        userId:
          type: string
          format: uuid
        userName:
          type: string
        active:
          type: boolean
    SheetListResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        currentPage:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        list:
          type: array
          items:
            $ref: '#/components/schemas/BasicSheetDto'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT